backend/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
yfinance>=0.2.28
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.2.0
matplotlib>=3.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
requests>=2.27.0
gunicorn>=20.1.0
